Use callback and implement watch handler at CLI level
